# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhisperBox API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: getMyMessages
      tags: [messages]
      summary: Get messages received by current user
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"
        "401":
          description: Unauthorized
  /messages/{username}:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send anonymous message to a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: User not found
        "429":
          description: Rate limit exceeded
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /messages/stats/me:
    get:
      operationId: getMyStats
      tags: [messages]
      summary: Get stats for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
        "401":
          description: Unauthorized
  /messages/wrapped/me:
    get:
      operationId: getMyWrapped
      tags: [messages]
      summary: Get Wrapped recap stats for a given period
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Wrapped stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WrappedStats"
        "401":
          description: Unauthorized
  /messages/{id}/reply:
    post:
      operationId: replyToMessage
      tags: [messages]
      summary: Reply to a message as the owner
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyMessageBody"
      responses:
        "200":
          description: Reply saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
        "404":
          description: Message not found
...
  /messages/{id}/visibility:
    patch:
      operationId: toggleMessageVisibility
      tags: [messages]
      summary: Toggle message public/private visibility
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisibilityBody"
      responses:
        "200":
          description: Visibility updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
        "404":
          description: Message not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        isRead:
          type: boolean
        isPublic:
          type: boolean
        ownerReply:
          type: string
          nullable: true
        ownerRepliedAt:
          type: string
          format: date-time
          nullable: true
        campaignTitle:
          type: string
          nullable: true
        senderEmail:
          type: string
          nullable: true
      required:
        - id
        - content
        - createdAt
        - isRead
        - isPublic
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - messages
        - total
        - page
        - limit
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
        senderEmail:
          type: string
          format: email
        turnstileToken:
          type: string
      required:
        - content
    ReplyMessageBody:
      type: object
      properties:
        reply:
          type: string
          minLength: 1
          maxLength: 1000
      required:
        - reply
    VisibilityBody:
      type: object
      properties:
        isPublic:
          type: boolean
      required:
        - isPublic
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    UserStats:
      type: object
      properties:
        totalMessages:
          type: integer
        unreadMessages:
          type: integer
        unrepliedMessages:
          type: integer
        linkOpens:
          type: integer
        isPremium:
          type: boolean
      required:
        - totalMessages
        - unreadMessages
        - unrepliedMessages
        - linkOpens
        - isPremium
    WrappedDayCount:
      type: object
      properties:
        day:
          type: integer
        count:
          type: integer
      required:
        - day
        - count
    WrappedHourCount:
      type: object
      properties:
        hour:
          type: integer
        count:
          type: integer
      required:
        - hour
        - count
    WrappedWordCount:
      type: object
      properties:
        word:
          type: string
        count:
          type: integer
      required:
        - word
        - count
    WrappedStats:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
          nullable: true
        totalMessages:
          type: integer
        repliedMessages:
          type: integer
        linkOpens:
          type: integer
        memberSince:
          type: string
          format: date-time
        dayDistribution:
          type: array
          items:
            $ref: "#/components/schemas/WrappedDayCount"
        hourDistribution:
          type: array
          items:
            $ref: "#/components/schemas/WrappedHourCount"
        longestMessage:
          type: string
          nullable: true
        topWords:
          type: array
          items:
            $ref: "#/components/schemas/WrappedWordCount"
      required:
        - year
        - totalMessages
        - repliedMessages
        - linkOpens
        - memberSince
        - dayDistribution
        - hourDistribution
        - topWords
...